requirements.txt
# ==========================================
# 🟢 GERAL (FUNCIONA EM MAC, LINUX E WINDOWS)
# Necessário para: Gerador de dados, manipulação de arquivos e utilitários
# ==========================================
datasets
pandas
scipy
sentencepiece
protobuf
tqdm
ollama
pypdf
python-dotenv

# No Mac, o torch é necessário para o VSCode não reclamar, 
# mas no Linux o Unsloth gerencia a versão dele.
torch
transformers

# Apenas para script do google colab
google

# ==========================================
# 🛑 DEPENDÊNCIAS EXCLUSIVAS NVIDIA (GPU / LINUX) 🛑
# ==========================================
# ATENÇÃO: Se estiver no Mac/Dev, COMENTE (adicione #) nas linhas abaixo.
# Estas libs exigem drivers CUDA que o Mac não possui.
# ==========================================

# unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
# --no-deps xformers<0.0.27
# --no-deps trl<0.9.0
# --no-deps peft
# --no-deps accelerate
# --no-deps bitsandbytes

# ==========================================
# 🛠 FERRAMENTAS DE DESENVOLVIMENTO
# ==========================================
pylint